feat: generic input/output add no_only_unit option
Without the REQKEY option, then
GENERIC_INPUT_FILE_INSTRUCTION and
GENERIC_OUTPUT_FILE_INSTRUCTION
will assume that an integer is a unit number
to read or write to.

This can cause problems if the input is regular
integer input, which is why the REQKEY option was added.

However, it may be desired to only require a keyword
for reading a unit number and not for specifying a
file name.

The option NO_ONLY_UNIT=.TRUE. allows for this.