refactor: num2str improved decimal support
NUM2STR(X), where X is a REAL32 or REAL64
now prints with 1, 3, or 5 decimal places
if that is all the number has.

For example, NUM2STR(3.1) outputs "3.1"

Previously, this had issues with certain decimals due to irrational binary numbers.
That is, 3.141 would output "3.10000" or "3.1e0"
due to the irrational decimal place in binary format.

The general code check was changed from:

   10*DVAL == AINT(10*DVAL)      ! -> Print with 1 decimal
to:
   ABS(10*DVAL - ANINT(10*DVAL)) < 10*DVAL * 1e-13