feat: generic_open subroutine argument to specify file open position
src/io/generic_open_interface.fpp subroutine generic_open
added the optional argument "position", which is then passed
to the Fortran open statement.

All Fortran OPEN(POSITION=) strings are accepted.
However it is recommended to only use:
"ASIS",   or
"REWIND", or
"APPEND"

If the POSITION argument is not given, then it is set to "REWIND".

This argument has no effect if the file is opened with
ACTION="WRITE" because the a new, empty file is created.