fix: added get_c_loc function in dynamic_array_int32
The C_ASSOCIATED and C_LOC functions from ISO_C_BINDING is used
to check if two allocatable arrays occupy the same memory space.
The recent compiler version now raises an error that C_LOC only
works with target or pointer arrays and not allocatable.

the function get_c_loc is added takes any array and sets the
target attribute and returns the result of C_LOC(array).

This replaces the following code:
    C_ASSOCIATED(C_LOC(dyn%array), C_LOC(dyn_in%array))
with
    C_ASSOCIATED(get_c_loc(dyn%array), get_c_loc(dyn_in%array))