refactor: change odd number check from iand to btest
The code previously used IAND to check if a number is odd
with the following code:

IAND(num,1) == 1

This bit folding only checks to see if the least
significant bit is set, which indicates the number is odd.

Rather than doing two operations, the code was changed to

BTEST(num, 0)

which returns TRUE if the lest significant bit is set.