fix: number_conversion_interface ensure consistent integer type
MODULE NUMBER_CONVERSION_INTERFACE converts all the major
integer types from/to binary to decimal, octal, or hexadecimal.
A set of parameters were added to ensure that zero and one
comparisons are made with the same type integer as that of the routine.
That is, converting INT16 to Octal uses only INT16 operations.

Also changed odd number checks from IAND to BTEST.