fix: change makefile back to ifort; gfortran has finalization errors
Unfortunetly after changing the makefile to use gfortran
as the default compiler was premature. After doing a
set of runtime tests, it was found that there are
allocation problems in the gfortran compiler
and it has issues with derived type final routines.

They have been submitted as compiler bugs to gcc.
Once they are fixed, gfortran will be added back as
as the default compiler.