fix: uload reads extra line for generic_input_file with specified unit
In the file src/input_reader/uload_and_sfac_interface.f90
the routine:
SUBROUTINE ULOAD_SCALAR

when loading a GENERIC_INPUT_FILE with a IU /= 0 (user supplied unit number)
resulted in ULOAD reading in a line to check for a scale factor (SFAC),
but failed to BACKSPACE to return the line to its original position.

This effected MODULE LIST_ARRAY_INPUT_INTERFACE when using
TYPE GENERIC_LINE_INPUT that reads a transient file reader (TFR) with
the DATAFILE or DATAUNIT keyword (bypass TFR and read input directly).
The resulting call to the MODULE TRANSIENT_FILE_READER_INSTRUCTION
should check for SFAC and set the GENERIC_INPUT_FILE to the
data file that is directly read from.