fix: uload fail to open file now prints correct error message
If ULOAD encounters a directive to access an external file
that does not exist resulted in an error message saying
that an implied internal read failed rather than file is missing.

This was corrected by adding to the GENERIC_INPUT object open call
the following option:
   KEY_FAIL_STOP=TRUE
which will stop the program if the file being accessed does not exist.