feat: ixj_instruction adds ixj_structure.get_value
The IXJ_INSTRUCTION contains the data type IXJ_STRUCTURE
that is a 1D array with each row containing a set of
integers, `I`, floats, `X`, and integers, `J`.
For example, a pumping well could be represented with
I = [ROW, COLUMN], X = [Q], and J = [].

This commit adds a GET_VALUE function that returns
the a value in X for a given set of I integers.