feat: compressed_value_storage type has separate capacity
The COMPRESSED_VALUE_STORAGE data type originally contained
a single dimension term `N` that represented the number of values.
Now, it contains an additional term `CAP` that is the size
of the allocated cvs storage arrays. This allows for the size
of the storage arrays to change with N, without requiring a reallocation,
so long as N <= CAP, otherwise a reallocate occurs.